2주차/mission2: 폴더 구조 변경
