+++ khaki/2주차/mission2/src/App.tsx
@@ -1,11 +1,14 @@
 import './App.css'
+import { ThemeProvider } from './useContext/context/ThemeProvider'
 import ContextPage from './useContext/ContextPage'
 
 
 const App = () => {
   return (
     <div>
-      <ContextPage/>
+      <ThemeProvider>
+        <ContextPage/>
+      </ThemeProvider>
     </div>
   )
 }
