+++ khaki/2주차/mission2/src/components/Navbar.tsx
@@ -4,7 +4,7 @@ import ThemeToggleButton from "./ThemeToggleButton";
 
 
 const Navbar = () => {
-  const {theme} = useTheme();
+  const {theme} = useTheme(); 
   const isLightMode = (theme === 'LIGHT');
