+++ khaki/2주차/mission2/src/components/ThemeContent.tsx
